Revert "Use `NoTextActionParameter`"
This reverts commit 4ca65f1b9220ca778d1a8051f3a296076644e34c.
