+++ sinks/dashboard/items/dashboard_item.py
@@ -7,8 +7,6 @@ from PySide6.QtCore import QRect
 import json
 from typing import Callable
 
-from .no_text_action_parameter import NoTextActionParameter
-
 
 class DashboardItem(QWidget):
     """
@@ -44,7 +42,7 @@ class DashboardItem(QWidget):
             {"name": "height", "type": "int", "value": 100}
         ])
 
-        lock_parameter = NoTextActionParameter(name="lock")
+        lock_parameter = ActionParameter(name="lock")
         lock_parameter.sigActivated.connect(self.lock)
         self.parameters.addChild(lock_parameter)
